Change vagrant box to eider/jessie64-vbox. Upgrade Vagrantfile and provision scripts

+++ provision/init.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 if [ "$#" -ne 1 ]; then
-  echo "Usage: $0 PROJECT_NAME" >&2
+  echo "Usage: \"$0\" project name" >&2
   exit 1
 fi
 export PROJECT_DIR=/var/www
@@ -23,8 +23,10 @@ echo "Reload Nginx service"
 systemctl restart nginx
 
 echo "Setup MySQL databases"
-mysql -u root -pvagrant -e "DROP USER '$PROJECT_NAME'@'%';"
+if [ `mysql -h localhost --port=11306 -u root -pvagrant mysql -e "SELECT COUNT(*) FROM user WHERE User = '$PROJECT_NAME' AND Host = '%'" -N -B` -gt 0 ]; then
+    mysql -u root -pvagrant -e "DROP USER '$PROJECT_NAME'@'%';"
+fi
 mysql -u root -pvagrant -e "CREATE USER '$PROJECT_NAME'@'%' IDENTIFIED BY '$PROJECT_NAME'; GRANT ALL ON \`$PROJECT_NAME%\`.* TO '$PROJECT_NAME'@'%'; FLUSH PRIVILEGES;"
-mysql -u root -pvagrant -e "DROP DATABASE IF EXISTS $PROJECT_NAME; CREATE DATABASE $PROJECT_NAME CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_unicode_ci';"
+mysql -u root -pvagrant -e "DROP DATABASE IF EXISTS \`$PROJECT_NAME\`; CREATE DATABASE \`$PROJECT_NAME\` CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_unicode_ci';"
 
 echo "==== Initializing finished"
